Submitting grades fixes #8
Users can now edit grades and submit new assignments into the system.
I've also changed the color scheme (again) to something that is somewhat
bearable but possibly too dark.

The class details page now shows the earned points and max points for
the class and the current letter grade, the same info that is posted on
the dashboard. The user doesn't have to go back to the dashboard to see
the letter grade anymore.

The iframe on the class details page will now resize to fit the contents
of the category details page. So, theoretically there should never be a
scrollbar in the iframe.