Admin Panel and a lot of other changes
There is now an admin control panel for users that have permissions set
to anything but 0. Working on creating reports so that team and league
leaders can view a breakdown of all user's grades and points.

I have also added more to the class category summary that includes the
real-time grade for that category. So even if there is only 1 assignment
in a category meant for many assignments the user can now see their real
time average and how that average is affecting their overall real time
grade.